packages/agent-worker/pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs", "grpcio-tools>=1.60.0"]
build-backend = "hatchling.build"

[project]
name = "agent-worker"
dynamic = ["version"]
description = "gRPC Agent SDK worker for claude-session-manager"
requires-python = ">=3.13"
dependencies = [
    "claude-agent-sdk>=0.1.0",
    "grpcio>=1.60.0",
    "protobuf>=4.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "grpcio-tools>=1.60.0",
    "ruff>=0.8.0",
]

[tool.hatch.version]
source = "vcs"
tag-pattern = "agent-worker/v(?P<version>.*)"
root = "../.."

[tool.hatch.build.hooks.custom]
path = "build_hook.py"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_worker"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
